@@ -1,12 +1,12 @@
 {
   "name": "@grpc/grpc-js",
-  "version": "0.3.6",
+  "version": "0.4.0",
   "description": "gRPC Library for Node - pure JS implementation",
   "homepage": "https://grpc.io/",
   "repository": "https://github.com/grpc/grpc-node/tree/master/packages/grpc-js",
   "main": "build/src/index.js",
   "engines": {
-    "node": "^8.11.2 || >=9.4"
+    "node": "^8.13.0 || >=10.10.0"
   },
   "keywords": [],
   "author": {
@@ -15,13 +15,14 @@
   "types": "build/src/index.d.ts",
   "license": "Apache-2.0",
   "devDependencies": {
+    "@grpc/proto-loader": "^0.5.0",
     "@types/lodash": "^4.14.108",
-    "@types/mocha": "^2.2.43",
-    "@types/node": "^10.5.4",
+    "@types/mocha": "^5.2.6",
+    "@types/node": "^12.0.2",
     "clang-format": "^1.0.55",
-    "gts": "^0.5.1",
+    "gts": "^1.0.0",
     "lodash": "^4.17.4",
-    "typescript": "~3.3.3333"
+    "typescript": "~3.4.5"
   },
   "contributors": [
     {
@@ -42,7 +43,7 @@
     "posttest": "npm run check"
   },
   "dependencies": {
-    "semver": "^5.5.0"
+    "semver": "^6.0.0"
   },
   "files": [
     "build/src/*.{js,d.ts}",
