@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 gRPC authors.
+ * Copyright 2019 gRPC authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,18 +15,13 @@
  *
  */
 
-const _gulp = require('gulp');
-const help = require('gulp-help');
-
-// gulp-help monkeypatches tasks to have an additional description parameter
-const gulp = help(_gulp);
-
-const jsdoc = require('gulp-jsdoc3');
-const jshint = require('gulp-jshint');
-const mocha = require('gulp-mocha');
-const execa = require('execa');
-const path = require('path');
-const del = require('del');
+ import * as gulp from 'gulp';
+ import * as jsdoc from 'gulp-jsdoc3';
+ import * as jshint from 'gulp-jshint';
+ import * as mocha from 'gulp-mocha';
+ import * as execa from 'execa';
+ import * as path from 'path';
+ import * as del from 'del';
 
 const nativeCoreDir = __dirname;
 const srcDir = path.resolve(nativeCoreDir, 'src');
@@ -35,44 +30,54 @@ const testDir = path.resolve(nativeCoreDir, 'test');
 const pkg = require('./package');
 const jshintConfig = pkg.jshintConfig;
 
-gulp.task('clean', 'Delete generated files', () => {
-  return del([path.resolve(nativeCoreDir, 'build'),
-	      path.resolve(nativeCoreDir, 'ext/node')]);
-});
+const clean = () => del([path.resolve(nativeCoreDir, 'build'),
+	                       path.resolve(nativeCoreDir, 'ext/node')]);
 
-gulp.task('clean.all', 'Delete all files created by tasks',
-	  ['clean']);
+const cleanAll = gulp.parallel(clean);
 
-gulp.task('install', 'Install native core dependencies', () => {
+const install = () => {
   return execa('npm', ['install', '--build-from-source', '--unsafe-perm'],
                {cwd: nativeCoreDir, stdio: 'inherit'});
-});
+};
 
-gulp.task('install.windows', 'Install native core dependencies for MS Windows', () => {
+const installWindows = () => {
   return execa('npm', ['install', '--build-from-source'],
                {cwd: nativeCoreDir, stdio: 'inherit'}).catch(() => 
 del(path.resolve(process.env.USERPROFILE, '.node-gyp', process.versions.node, 'include/node/openssl'), { force: true }).then(() =>
 execa('npm', ['install', '--build-from-source'],
                {cwd: nativeCoreDir, stdio: 'inherit'})
-               ))
-});
+               ));
+};
 
-gulp.task('lint', 'Emits linting errors', () => {
+const lint = () => {
   return gulp.src([`${nativeCoreDir}/index.js`, `${srcDir}/*.js`, `${testDir}/*.js`])
       .pipe(jshint(pkg.jshintConfig))
       .pipe(jshint.reporter('default'));
-});
+};
 
-gulp.task('build', 'Build native package', () => {
+const build = () => {
   return execa('npm', ['run', 'build'], {cwd: nativeCoreDir, stdio: 'inherit'});
-});
+};
 
-gulp.task('test', 'Run all tests', ['build'], () => {
+const runTests = () => {
   return gulp.src(`${testDir}/*.js`).pipe(mocha({timeout: 5000, reporter: 'mocha-jenkins-reporter'}));
-});
+}
 
-gulp.task('doc.gen', 'Generate docs', (cb) => {
+const test = gulp.series(build, runTests);
+
+const docGen = (cb) => {
   var config = require('./jsdoc_conf.json');
-  gulp.src([`${nativeCoreDir}/README.md`, `${nativeCoreDir}/index.js`, `${srcDir}/*.js`], {read: false})
-      .pipe(jsdoc(config, cb));
-});
+  return gulp.src([`${nativeCoreDir}/README.md`, `${nativeCoreDir}/index.js`, `${srcDir}/*.js`], {read: false})
+             .pipe(jsdoc(config, cb));
+};
+
+export {
+  clean,
+  cleanAll,
+  install,
+  installWindows,
+  lint,
+  build,
+  test,
+  docGen
+};
