@@ -15,10 +15,8 @@
  *
  */
 
-import * as _gulp from 'gulp';
-import * as help from 'gulp-help';
+import * as gulp from 'gulp';
 
-import * as fs from 'fs';
 import * as mocha from 'gulp-mocha';
 import * as path from 'path';
 import * as execa from 'execa';
@@ -26,58 +24,61 @@ import * as pify from 'pify';
 import * as semver from 'semver';
 import { ncp } from 'ncp';
 
-// gulp-help monkeypatches tasks to have an additional description parameter
-const gulp = help(_gulp);
-
 const ncpP = pify(ncp);
 
 Error.stackTraceLimit = Infinity;
 
 const jsCoreDir = __dirname;
-const tslintPath = path.resolve(jsCoreDir, 'node_modules/google-ts-style/tslint.json');
-const tsconfigPath = path.resolve(jsCoreDir, 'tsconfig.json');
 const outDir = path.resolve(jsCoreDir, 'build');
-const srcDir = path.resolve(jsCoreDir, 'src');
-const testDir = path.resolve(jsCoreDir, 'test');
+
+const pkgPath = path.resolve(jsCoreDir, 'package.json');
+const supportedVersionRange = require(pkgPath).engines.node;
+const versionNotSupported = () => {
+  console.log(`Skipping grpc-js task for Node ${process.version}`);
+  return () => { return Promise.resolve(); };
+};
+const identity = (value: any): any => value;
+const checkTask = semver.satisfies(process.version, supportedVersionRange) ?
+    identity : versionNotSupported;
 
 const execNpmVerb = (verb: string, ...args: string[]) =>
   execa('npm', [verb, ...args], {cwd: jsCoreDir, stdio: 'inherit'});
 const execNpmCommand = execNpmVerb.bind(null, 'run');
 
-gulp.task('install', 'Install native core dependencies', () =>
-  execNpmVerb('install', '--unsafe-perm'));
+const install = checkTask(() => execNpmVerb('install', '--unsafe-perm'));
 
 /**
  * Runs tslint on files in src/, with linting rules defined in tslint.json.
  */
-gulp.task('lint', 'Emits linting errors found in src/ and test/.', () =>
-  execNpmCommand('check'));
+const lint = checkTask(() => execNpmCommand('check'));
 
-gulp.task('clean', 'Deletes transpiled code.', ['install'],
-  () => execNpmCommand('clean'));
+const cleanFiles = checkTask(() => execNpmCommand('clean'));
 
-gulp.task('clean.all', 'Deletes all files added by targets', ['clean']);
+const clean = gulp.series(install, cleanFiles);
+
+const cleanAll = gulp.parallel(clean);
 
 /**
  * Transpiles TypeScript files in src/ to JavaScript according to the settings
  * found in tsconfig.json.
  */
-gulp.task('compile', 'Transpiles src/.', () => execNpmCommand('compile'));
+const compile = checkTask(() => execNpmCommand('compile'));
 
-gulp.task('copy-test-fixtures', 'Copy test fixtures.', () => {
-  return ncpP(`${jsCoreDir}/test/fixtures`, `${outDir}/test/fixtures`);
+const copyTestFixtures = checkTask(() => ncpP(`${jsCoreDir}/test/fixtures`, `${outDir}/test/fixtures`));
+
+const runTests = checkTask(() => {
+  return gulp.src(`${outDir}/test/**/*.js`)
+    .pipe(mocha({reporter: 'mocha-jenkins-reporter',
+                 require: ['ts-node/register']}));
 });
 
-/**
- * Transpiles src/ and test/, and then runs all tests.
- */
-gulp.task('test', 'Runs all tests.', ['lint', 'copy-test-fixtures'], () => {
-  if (semver.satisfies(process.version, '^8.11.2 || >=9.4')) {
-    return gulp.src(`${outDir}/test/**/*.js`)
-      .pipe(mocha({reporter: 'mocha-jenkins-reporter',
-                   require: ['ts-node/register']}));
-  } else {
-    console.log(`Skipping grpc-js tests for Node ${process.version}`);
-    return Promise.resolve(null);
-  }
-});
+const test = gulp.series(install, copyTestFixtures, runTests);
+
+export {
+  install,
+  lint,
+  clean,
+  cleanAll,
+  compile,
+  test
+}
