@@ -1,13 +1,32 @@
-import {CallCredentials} from './call-credentials';
-import {Call} from './call-stream';
-import {Http2Channel} from './channel';
-import {BaseFilter, Filter, FilterFactory} from './filter';
-import {Metadata} from './metadata';
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import { CallCredentials } from './call-credentials';
+import { Call } from './call-stream';
+import { Http2Channel } from './channel';
+import { BaseFilter, Filter, FilterFactory } from './filter';
+import { Metadata } from './metadata';
 
 export class CallCredentialsFilter extends BaseFilter implements Filter {
   private serviceUrl: string;
   constructor(
-      private readonly channel: Http2Channel, private readonly stream: Call) {
+    private readonly channel: Http2Channel,
+    private readonly stream: Call
+  ) {
     super();
     this.channel = channel;
     this.stream = stream;
@@ -28,16 +47,17 @@ export class CallCredentialsFilter extends BaseFilter implements Filter {
     const channelCredentials = this.channel.credentials._getCallCredentials();
     const streamCredentials = this.stream.getCredentials();
     const credentials = channelCredentials.compose(streamCredentials);
-    const credsMetadata =
-        credentials.generateMetadata({service_url: this.serviceUrl});
+    const credsMetadata = credentials.generateMetadata({
+      service_url: this.serviceUrl,
+    });
     const resultMetadata = await metadata;
     resultMetadata.merge(await credsMetadata);
     return resultMetadata;
   }
 }
 
-export class CallCredentialsFilterFactory implements
-    FilterFactory<CallCredentialsFilter> {
+export class CallCredentialsFilterFactory
+  implements FilterFactory<CallCredentialsFilter> {
   private readonly channel: Http2Channel;
   constructor(channel: Http2Channel) {
     this.channel = channel;
