@@ -1,25 +1,46 @@
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
 import * as http2 from 'http2';
-import {Duplex} from 'stream';
+import { Duplex } from 'stream';
 
-import {CallCredentials} from './call-credentials';
-import {Http2Channel} from './channel';
-import {Status} from './constants';
-import {EmitterAugmentation1} from './events';
-import {Filter} from './filter';
-import {FilterStackFactory} from './filter-stack';
-import {Metadata} from './metadata';
-import {ObjectDuplex, WriteCallback} from './object-stream';
+import { CallCredentials } from './call-credentials';
+import { Http2Channel } from './channel';
+import { Status } from './constants';
+import { EmitterAugmentation1 } from './events';
+import { Filter } from './filter';
+import { FilterStackFactory } from './filter-stack';
+import { Metadata } from './metadata';
+import { ObjectDuplex, WriteCallback } from './object-stream';
+import { StreamDecoder } from './stream-decoder';
 
-const {HTTP2_HEADER_STATUS, HTTP2_HEADER_CONTENT_TYPE, NGHTTP2_CANCEL} =
-    http2.constants;
+const {
+  HTTP2_HEADER_STATUS,
+  HTTP2_HEADER_CONTENT_TYPE,
+  NGHTTP2_CANCEL,
+} = http2.constants;
 
-export type Deadline = Date|number;
+export type Deadline = Date | number;
 
 export interface CallStreamOptions {
   deadline: Deadline;
   flags: number;
   host: string;
-  parentCall: Call|null;
+  parentCall: Call | null;
 }
 
 export type PartialCallStreamOptions = Partial<CallStreamOptions>;
@@ -33,7 +54,7 @@ export interface StatusObject {
 export const enum WriteFlags {
   BufferHint = 1,
   NoCompress = 2,
-  WriteThrough = 4
+  WriteThrough = 4,
 }
 
 export interface WriteObject {
@@ -45,7 +66,8 @@ export interface WriteObject {
  * This interface represents a duplex stream associated with a single gRPC call.
  */
 export type Call = {
-  cancelWithStatus(status: Status, details: string): void; getPeer(): string;
+  cancelWithStatus(status: Status, details: string): void;
+  getPeer(): string;
   sendMetadata(metadata: Metadata): void;
 
   getDeadline(): Deadline;
@@ -56,38 +78,26 @@ export type Call = {
   getStatus(): StatusObject | null;
   getMethod(): string;
   getHost(): string;
-}&EmitterAugmentation1<'metadata', Metadata>&
-    EmitterAugmentation1<'status', StatusObject>&
-    ObjectDuplex<WriteObject, Buffer>;
-
-enum ReadState {
-  NO_DATA,
-  READING_SIZE,
-  READING_MESSAGE
-}
+} & EmitterAugmentation1<'metadata', Metadata> &
+  EmitterAugmentation1<'status', StatusObject> &
+  ObjectDuplex<WriteObject, Buffer>;
 
 export class Http2CallStream extends Duplex implements Call {
   credentials: CallCredentials = CallCredentials.createEmpty();
   filterStack: Filter;
-  private http2Stream: http2.ClientHttp2Stream|null = null;
+  private http2Stream: http2.ClientHttp2Stream | null = null;
   private pendingRead = false;
-  private pendingWrite: Buffer|null = null;
-  private pendingWriteCallback: WriteCallback|null = null;
-  private pendingFinalCallback: Function|null = null;
+  private pendingWrite: Buffer | null = null;
+  private pendingWriteCallback: WriteCallback | null = null;
+  private pendingFinalCallback: Function | null = null;
 
-  private readState: ReadState = ReadState.NO_DATA;
-  private readCompressFlag: Buffer = Buffer.alloc(1);
-  private readPartialSize: Buffer = Buffer.alloc(4);
-  private readSizeRemaining = 4;
-  private readMessageSize = 0;
-  private readPartialMessage: Buffer[] = [];
-  private readMessageRemaining = 0;
+  private decoder = new StreamDecoder();
 
   private isReadFilterPending = false;
   private canPush = false;
 
-  private unpushedReadMessages: Array<Buffer|null> = [];
-  private unfilteredReadMessages: Array<Buffer|null> = [];
+  private unpushedReadMessages: Array<Buffer | null> = [];
+  private unfilteredReadMessages: Array<Buffer | null> = [];
 
   // Status code mapped from :status. To be used if grpc-status is not received
   private mappedStatusCode: Status = Status.UNKNOWN;
@@ -100,14 +110,15 @@ export class Http2CallStream extends Duplex implements Call {
   private handlingTrailers = Promise.resolve();
 
   // This is populated (non-null) if and only if the call has ended
-  private finalStatus: StatusObject|null = null;
+  private finalStatus: StatusObject | null = null;
 
   constructor(
-      private readonly methodName: string,
-      private readonly channel: Http2Channel,
-      private readonly options: CallStreamOptions,
-      filterStackFactory: FilterStackFactory) {
-    super({objectMode: true});
+    private readonly methodName: string,
+    private readonly channel: Http2Channel,
+    private readonly options: CallStreamOptions,
+    filterStackFactory: FilterStackFactory
+  ) {
+    super({ objectMode: true });
     this.filterStack = filterStackFactory.createFilter(this);
   }
 
@@ -150,7 +161,7 @@ export class Http2CallStream extends Duplex implements Call {
     }
   }
 
-  private filterReceivedMessage(framedMessage: Buffer|null) {
+  private filterReceivedMessage(framedMessage: Buffer | null) {
     /* If we the call has already ended, we don't want to do anything with
      * this message. Dropping it on the floor is correct behavior */
     if (this.finalStatus !== null) {
@@ -165,13 +176,15 @@ export class Http2CallStream extends Duplex implements Call {
       return;
     }
     this.isReadFilterPending = true;
-    this.filterStack.receiveMessage(Promise.resolve(framedMessage))
-        .then(
-            this.handleFilteredRead.bind(this),
-            this.handleFilterError.bind(this));
+    this.filterStack
+      .receiveMessage(Promise.resolve(framedMessage))
+      .then(
+        this.handleFilteredRead.bind(this),
+        this.handleFilterError.bind(this)
+      );
   }
 
-  private tryPush(messageBytes: Buffer|null): void {
+  private tryPush(messageBytes: Buffer | null): void {
     if (this.isReadFilterPending) {
       this.unfilteredReadMessages.push(messageBytes);
     } else {
@@ -188,20 +201,21 @@ export class Http2CallStream extends Duplex implements Call {
     } catch (e) {
       metadata = new Metadata();
     }
-    const status: StatusObject = {code, details, metadata};
+    const status: StatusObject = { code, details, metadata };
     this.handlingTrailers = (async () => {
       let finalStatus;
       try {
         // Attempt to assign final status.
-        finalStatus =
-            await this.filterStack.receiveTrailers(Promise.resolve(status));
+        finalStatus = await this.filterStack.receiveTrailers(
+          Promise.resolve(status)
+        );
       } catch (error) {
         await this.handlingHeaders;
         // This is a no-op if the call was already ended when handling headers.
         this.endCall({
           code: Status.INTERNAL,
           details: 'Failed to process received status',
-          metadata: new Metadata()
+          metadata: new Metadata(),
         });
         return;
       }
@@ -254,89 +268,37 @@ export class Http2CallStream extends Duplex implements Call {
             this.endCall({
               code: Status.UNKNOWN,
               details: error.message,
-              metadata: new Metadata()
+              metadata: new Metadata(),
             });
             return;
           }
-          this.handlingHeaders =
-              this.filterStack.receiveMetadata(Promise.resolve(metadata))
-                  .then((finalMetadata) => {
-                    this.emit('metadata', finalMetadata);
-                  })
-                  .catch((error) => {
-                    this.destroyHttp2Stream();
-                    this.endCall({
-                      code: Status.UNKNOWN,
-                      details: error.message,
-                      metadata: new Metadata()
-                    });
-                  });
+          this.handlingHeaders = this.filterStack
+            .receiveMetadata(Promise.resolve(metadata))
+            .then(finalMetadata => {
+              this.emit('metadata', finalMetadata);
+            })
+            .catch(error => {
+              this.destroyHttp2Stream();
+              this.endCall({
+                code: Status.UNKNOWN,
+                details: error.message,
+                metadata: new Metadata(),
+              });
+            });
         }
       });
       stream.on('trailers', this.handleTrailers.bind(this));
       stream.on('data', (data: Buffer) => {
-        let readHead = 0;
-        let toRead: number;
-        while (readHead < data.length) {
-          switch (this.readState) {
-            case ReadState.NO_DATA:
-              this.readCompressFlag = data.slice(readHead, readHead + 1);
-              readHead += 1;
-              this.readState = ReadState.READING_SIZE;
-              this.readPartialSize.fill(0);
-              this.readSizeRemaining = 4;
-              this.readMessageSize = 0;
-              this.readMessageRemaining = 0;
-              this.readPartialMessage = [];
-              break;
-            case ReadState.READING_SIZE:
-              toRead = Math.min(data.length - readHead, this.readSizeRemaining);
-              data.copy(
-                  this.readPartialSize, 4 - this.readSizeRemaining, readHead,
-                  readHead + toRead);
-              this.readSizeRemaining -= toRead;
-              readHead += toRead;
-              // readSizeRemaining >=0 here
-              if (this.readSizeRemaining === 0) {
-                this.readMessageSize = this.readPartialSize.readUInt32BE(0);
-                this.readMessageRemaining = this.readMessageSize;
-                if (this.readMessageRemaining > 0) {
-                  this.readState = ReadState.READING_MESSAGE;
-                } else {
-                  this.tryPush(Buffer.concat(
-                      [this.readCompressFlag, this.readPartialSize]));
-                  this.readState = ReadState.NO_DATA;
-                }
-              }
-              break;
-            case ReadState.READING_MESSAGE:
-              toRead =
-                  Math.min(data.length - readHead, this.readMessageRemaining);
-              this.readPartialMessage.push(
-                  data.slice(readHead, readHead + toRead));
-              this.readMessageRemaining -= toRead;
-              readHead += toRead;
-              // readMessageRemaining >=0 here
-              if (this.readMessageRemaining === 0) {
-                // At this point, we have read a full message
-                const framedMessageBuffers = [
-                  this.readCompressFlag, this.readPartialSize
-                ].concat(this.readPartialMessage);
-                const framedMessage = Buffer.concat(
-                    framedMessageBuffers, this.readMessageSize + 5);
-                this.tryPush(framedMessage);
-                this.readState = ReadState.NO_DATA;
-              }
-              break;
-            default:
-              throw new Error('This should never happen');
-          }
+        const message = this.decoder.write(data);
+
+        if (message !== null) {
+          this.tryPush(message);
         }
       });
       stream.on('end', () => {
         this.tryPush(null);
       });
-      stream.on('close', async (errorCode) => {
+      stream.on('close', async errorCode => {
         let code: Status;
         let details = '';
         switch (errorCode) {
@@ -364,13 +326,13 @@ export class Http2CallStream extends Duplex implements Call {
         // This is OK, because status codes emitted here correspond to more
         // catastrophic issues that prevent us from receiving trailers in the
         // first place.
-        this.endCall({code, details, metadata: new Metadata()});
+        this.endCall({ code, details, metadata: new Metadata() });
       });
       stream.on('error', (err: Error) => {
         this.endCall({
           code: Status.INTERNAL,
           details: 'Internal HTTP2 error',
-          metadata: new Metadata()
+          metadata: new Metadata(),
         });
       });
       if (!this.pendingRead) {
@@ -390,7 +352,11 @@ export class Http2CallStream extends Duplex implements Call {
 
   sendMetadata(metadata: Metadata): void {
     this.channel._startHttp2Stream(
-        this.options.host, this.methodName, this, metadata);
+      this.options.host,
+      this.methodName,
+      this,
+      metadata
+    );
   }
 
   private destroyHttp2Stream() {
@@ -409,7 +375,7 @@ export class Http2CallStream extends Duplex implements Call {
       // If trailers are currently being processed, the call should be ended
       // by handleTrailers instead.
       await this.handlingTrailers;
-      this.endCall({code: status, details, metadata: new Metadata()});
+      this.endCall({ code: status, details, metadata: new Metadata() });
     })();
   }
 
@@ -425,7 +391,7 @@ export class Http2CallStream extends Duplex implements Call {
     this.credentials = credentials;
   }
 
-  getStatus(): StatusObject|null {
+  getStatus(): StatusObject | null {
     return this.finalStatus;
   }
 
@@ -455,7 +421,7 @@ export class Http2CallStream extends Duplex implements Call {
       while (this.unpushedReadMessages.length > 0) {
         const nextMessage = this.unpushedReadMessages.shift();
         this.canPush = this.push(nextMessage);
-        if (nextMessage === null || (!this.canPush)) {
+        if (nextMessage === null || !this.canPush) {
           this.canPush = false;
           return;
         }
@@ -468,7 +434,7 @@ export class Http2CallStream extends Duplex implements Call {
   }
 
   _write(chunk: WriteObject, encoding: string, cb: WriteCallback) {
-    this.filterStack.sendMessage(Promise.resolve(chunk)).then((message) => {
+    this.filterStack.sendMessage(Promise.resolve(chunk)).then(message => {
       if (this.http2Stream === null) {
         this.pendingWrite = message.message;
         this.pendingWriteCallback = cb;
