@@ -1,24 +1,43 @@
-import {EventEmitter} from 'events';
-import {Duplex, Readable, Writable} from 'stream';
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
 
-import {Call, StatusObject, WriteObject} from './call-stream';
-import {Status} from './constants';
-import {EmitterAugmentation1} from './events';
-import {Metadata} from './metadata';
-import {ObjectReadable, ObjectWritable} from './object-stream';
+import { EventEmitter } from 'events';
+import { Duplex, Readable, Writable } from 'stream';
+
+import { Call, StatusObject, WriteObject } from './call-stream';
+import { Status } from './constants';
+import { EmitterAugmentation1 } from './events';
+import { Metadata } from './metadata';
+import { ObjectReadable, ObjectWritable } from './object-stream';
 
 /**
  * A type extending the built-in Error object with additional fields.
  */
-export type ServiceError = StatusObject&Error;
+export type ServiceError = StatusObject & Error;
 
 /**
  * A base type for all user-facing values returned by client-side method calls.
  */
 export type SurfaceCall = {
-  cancel(): void; getPeer(): string;
-}&EmitterAugmentation1<'metadata', Metadata>&
-    EmitterAugmentation1<'status', StatusObject>&EventEmitter;
+  cancel(): void;
+  getPeer(): string;
+} & EmitterAugmentation1<'metadata', Metadata> &
+  EmitterAugmentation1<'status', StatusObject> &
+  EventEmitter;
 
 /**
  * A type representing the return value of a unary method call.
@@ -30,23 +49,27 @@ export type ClientUnaryCall = SurfaceCall;
  */
 export type ClientReadableStream<ResponseType> = {
   deserialize: (chunk: Buffer) => ResponseType;
-}&SurfaceCall&ObjectReadable<ResponseType>;
+} & SurfaceCall &
+  ObjectReadable<ResponseType>;
 
 /**
  * A type representing the return value of a client stream method call.
  */
 export type ClientWritableStream<RequestType> = {
   serialize: (value: RequestType) => Buffer;
-}&SurfaceCall&ObjectWritable<RequestType>;
+} & SurfaceCall &
+  ObjectWritable<RequestType>;
 
 /**
  * A type representing the return value of a bidirectional stream method call.
  */
-export type ClientDuplexStream<RequestType, ResponseType> =
-    ClientWritableStream<RequestType>&ClientReadableStream<ResponseType>;
+export type ClientDuplexStream<
+  RequestType,
+  ResponseType
+> = ClientWritableStream<RequestType> & ClientReadableStream<ResponseType>;
 
-export class ClientUnaryCallImpl extends EventEmitter implements
-    ClientUnaryCall {
+export class ClientUnaryCallImpl extends EventEmitter
+  implements ClientUnaryCall {
   constructor(private readonly call: Call) {
     super();
     call.on('metadata', (metadata: Metadata) => {
@@ -67,8 +90,10 @@ export class ClientUnaryCallImpl extends EventEmitter implements
 }
 
 function setUpReadableStream<ResponseType>(
-    stream: ClientReadableStream<ResponseType>, call: Call,
-    deserialize: (chunk: Buffer) => ResponseType): void {
+  stream: ClientReadableStream<ResponseType>,
+  call: Call,
+  deserialize: (chunk: Buffer) => ResponseType
+): void {
   let statusEmitted = false;
   call.on('data', (data: Buffer) => {
     let deserialized: ResponseType;
@@ -93,8 +118,10 @@ function setUpReadableStream<ResponseType>(
   });
   call.on('status', (status: StatusObject) => {
     if (status.code !== Status.OK) {
-      const error: ServiceError =
-          Object.assign(new Error(status.details), status);
+      const error: ServiceError = Object.assign(
+        new Error(status.details),
+        status
+      );
       stream.emit('error', error);
     }
     stream.emit('status', status);
@@ -103,12 +130,13 @@ function setUpReadableStream<ResponseType>(
   call.pause();
 }
 
-export class ClientReadableStreamImpl<ResponseType> extends Readable implements
-    ClientReadableStream<ResponseType> {
+export class ClientReadableStreamImpl<ResponseType> extends Readable
+  implements ClientReadableStream<ResponseType> {
   constructor(
-      private readonly call: Call,
-      readonly deserialize: (chunk: Buffer) => ResponseType) {
-    super({objectMode: true});
+    private readonly call: Call,
+    readonly deserialize: (chunk: Buffer) => ResponseType
+  ) {
+    super({ objectMode: true });
     call.on('metadata', (metadata: Metadata) => {
       this.emit('metadata', metadata);
     });
@@ -129,8 +157,12 @@ export class ClientReadableStreamImpl<ResponseType> extends Readable implements
 }
 
 function tryWrite<RequestType>(
-    call: Call, serialize: (value: RequestType) => Buffer, chunk: RequestType,
-    encoding: string, cb: Function) {
+  call: Call,
+  serialize: (value: RequestType) => Buffer,
+  chunk: RequestType,
+  encoding: string,
+  cb: Function
+) {
   let message: Buffer;
   const flags: number = Number(encoding);
   try {
@@ -140,19 +172,20 @@ function tryWrite<RequestType>(
     cb(e);
     return;
   }
-  const writeObj: WriteObject = {message};
+  const writeObj: WriteObject = { message };
   if (!Number.isNaN(flags)) {
     writeObj.flags = flags;
   }
   call.write(writeObj, cb);
 }
 
-export class ClientWritableStreamImpl<RequestType> extends Writable implements
-    ClientWritableStream<RequestType> {
+export class ClientWritableStreamImpl<RequestType> extends Writable
+  implements ClientWritableStream<RequestType> {
   constructor(
-      private readonly call: Call,
-      readonly serialize: (value: RequestType) => Buffer) {
-    super({objectMode: true});
+    private readonly call: Call,
+    readonly serialize: (value: RequestType) => Buffer
+  ) {
+    super({ objectMode: true });
     call.on('metadata', (metadata: Metadata) => {
       this.emit('metadata', metadata);
     });
@@ -180,12 +213,13 @@ export class ClientWritableStreamImpl<RequestType> extends Writable implements
 }
 
 export class ClientDuplexStreamImpl<RequestType, ResponseType> extends Duplex
-    implements ClientDuplexStream<RequestType, ResponseType> {
+  implements ClientDuplexStream<RequestType, ResponseType> {
   constructor(
-      private readonly call: Call,
-      readonly serialize: (value: RequestType) => Buffer,
-      readonly deserialize: (chunk: Buffer) => ResponseType) {
-    super({objectMode: true});
+    private readonly call: Call,
+    readonly serialize: (value: RequestType) => Buffer,
+    readonly deserialize: (chunk: Buffer) => ResponseType
+  ) {
+    super({ objectMode: true });
     call.on('metadata', (metadata: Metadata) => {
       this.emit('metadata', metadata);
     });
