@@ -1,21 +1,43 @@
-import {EventEmitter} from 'events';
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import { EventEmitter } from 'events';
 import * as http2 from 'http2';
-import {checkServerIdentity, PeerCertificate} from 'tls';
+import { checkServerIdentity, PeerCertificate } from 'tls';
 import * as url from 'url';
 
-import {CallCredentialsFilterFactory} from './call-credentials-filter';
-import {Call, CallStreamOptions, Deadline, Http2CallStream} from './call-stream';
-import {ChannelCredentials} from './channel-credentials';
-import {ChannelOptions, recognizedOptions} from './channel-options';
-import {CompressionFilterFactory} from './compression-filter';
-import {Status} from './constants';
-import {DeadlineFilterFactory} from './deadline-filter';
-import {FilterStackFactory} from './filter-stack';
-import {Metadata} from './metadata';
-import {MetadataStatusFilterFactory} from './metadata-status-filter';
-import {Http2SubChannel} from './subchannel';
+import { CallCredentialsFilterFactory } from './call-credentials-filter';
+import {
+  Call,
+  CallStreamOptions,
+  Deadline,
+  Http2CallStream,
+} from './call-stream';
+import { ChannelCredentials } from './channel-credentials';
+import { ChannelOptions, recognizedOptions } from './channel-options';
+import { CompressionFilterFactory } from './compression-filter';
+import { Status } from './constants';
+import { DeadlineFilterFactory } from './deadline-filter';
+import { FilterStackFactory } from './filter-stack';
+import { Metadata } from './metadata';
+import { MetadataStatusFilterFactory } from './metadata-status-filter';
+import { Http2SubChannel } from './subchannel';
 
-const {version: clientVersion} = require('../../package.json');
+const { version: clientVersion } = require('../../package.json');
 
 const MIN_CONNECT_TIMEOUT_MS = 20000;
 const INITIAL_BACKOFF_MS = 1000;
@@ -29,7 +51,7 @@ const {
   HTTP2_HEADER_METHOD,
   HTTP2_HEADER_PATH,
   HTTP2_HEADER_TE,
-  HTTP2_HEADER_USER_AGENT
+  HTTP2_HEADER_USER_AGENT,
 } = http2.constants;
 
 export enum ConnectivityState {
@@ -37,7 +59,7 @@ export enum ConnectivityState {
   READY,
   TRANSIENT_FAILURE,
   IDLE,
-  SHUTDOWN
+  SHUTDOWN,
 }
 
 function uniformRandom(min: number, max: number) {
@@ -81,8 +103,10 @@ export interface Channel {
    *     error if the deadline passes without a state change.
    */
   watchConnectivityState(
-      currentState: ConnectivityState, deadline: Date|number,
-      callback: (error?: Error) => void): void;
+    currentState: ConnectivityState,
+    deadline: Date | number,
+    callback: (error?: Error) => void
+  ): void;
   /**
    * Create a call object. Call is an opaque type that is used by the Client
    * class. This function is called by the gRPC library when starting a
@@ -96,9 +120,12 @@ export interface Channel {
    *     that indicates what information to propagate from parentCall.
    */
   createCall(
-      method: string, deadline: Deadline|null|undefined,
-      host: string|null|undefined, parentCall: Call|null|undefined,
-      propagateFlags: number|null|undefined): Call;
+    method: string,
+    deadline: Deadline | null | undefined,
+    host: string | null | undefined,
+    parentCall: Call | null | undefined,
+    propagateFlags: number | null | undefined
+  ): Call;
 }
 
 export class Http2Channel extends EventEmitter implements Channel {
@@ -107,10 +134,10 @@ export class Http2Channel extends EventEmitter implements Channel {
   private readonly defaultAuthority: string;
   private connectivityState: ConnectivityState = ConnectivityState.IDLE;
   // Helper Promise object only used in the implementation of connect().
-  private connecting: Promise<void>|null = null;
+  private connecting: Promise<void> | null = null;
   /* For now, we have up to one subchannel, which will exist as long as we are
    * connecting or trying to connect */
-  private subChannel: Http2SubChannel|null = null;
+  private subChannel: Http2SubChannel | null = null;
   private filterStackFactory: FilterStackFactory;
 
   private subChannelConnectCallback: () => void = () => {};
@@ -121,21 +148,28 @@ export class Http2Channel extends EventEmitter implements Channel {
   private currentBackoffDeadline: Date;
 
   private handleStateChange(
-      oldState: ConnectivityState, newState: ConnectivityState): void {
+    oldState: ConnectivityState,
+    newState: ConnectivityState
+  ): void {
     const now: Date = new Date();
     switch (newState) {
       case ConnectivityState.CONNECTING:
         if (oldState === ConnectivityState.IDLE) {
           this.currentBackoff = INITIAL_BACKOFF_MS;
-          this.currentBackoffDeadline =
-              new Date(now.getTime() + INITIAL_BACKOFF_MS);
+          this.currentBackoffDeadline = new Date(
+            now.getTime() + INITIAL_BACKOFF_MS
+          );
         } else if (oldState === ConnectivityState.TRANSIENT_FAILURE) {
           this.currentBackoff = Math.min(
-              this.currentBackoff * BACKOFF_MULTIPLIER, MAX_BACKOFF_MS);
+            this.currentBackoff * BACKOFF_MULTIPLIER,
+            MAX_BACKOFF_MS
+          );
           const jitterMagnitude: number = BACKOFF_JITTER * this.currentBackoff;
           this.currentBackoffDeadline = new Date(
-              now.getTime() + this.currentBackoff +
-              uniformRandom(-jitterMagnitude, jitterMagnitude));
+            now.getTime() +
+              this.currentBackoff +
+              uniformRandom(-jitterMagnitude, jitterMagnitude)
+          );
         }
         this.startConnecting();
         break;
@@ -146,8 +180,9 @@ export class Http2Channel extends EventEmitter implements Channel {
         this.subChannel = null;
         this.backoffTimerId = setTimeout(() => {
           this.transitionToState(
-              [ConnectivityState.TRANSIENT_FAILURE],
-              ConnectivityState.CONNECTING);
+            [ConnectivityState.TRANSIENT_FAILURE],
+            ConnectivityState.CONNECTING
+          );
         }, this.currentBackoffDeadline.getTime() - now.getTime());
         break;
       case ConnectivityState.IDLE:
@@ -155,7 +190,9 @@ export class Http2Channel extends EventEmitter implements Channel {
         if (this.subChannel) {
           this.subChannel.close();
           this.subChannel.removeListener(
-              'connect', this.subChannelConnectCallback);
+            'connect',
+            this.subChannelConnectCallback
+          );
           this.subChannel.removeListener('close', this.subChannelCloseCallback);
           this.subChannel = null;
           this.emit('shutdown');
@@ -169,7 +206,9 @@ export class Http2Channel extends EventEmitter implements Channel {
 
   // Transition from any of a set of oldStates to a specific newState
   private transitionToState(
-      oldStates: ConnectivityState[], newState: ConnectivityState): void {
+    oldStates: ConnectivityState[],
+    newState: ConnectivityState
+  ): void {
     if (oldStates.indexOf(this.connectivityState) > -1) {
       const oldState: ConnectivityState = this.connectivityState;
       this.connectivityState = newState;
@@ -180,28 +219,36 @@ export class Http2Channel extends EventEmitter implements Channel {
 
   private startConnecting(): void {
     const connectionOptions: http2.SecureClientSessionOptions =
-        this.credentials._getConnectionOptions() || {};
+      this.credentials._getConnectionOptions() || {};
     if (connectionOptions.secureContext !== null) {
       // If provided, the value of grpc.ssl_target_name_override should be used
       // to override the target hostname when checking server identity.
       // This option is used for testing only.
       if (this.options['grpc.ssl_target_name_override']) {
-        const sslTargetNameOverride =
-            this.options['grpc.ssl_target_name_override']!;
-        connectionOptions.checkServerIdentity =
-            (host: string, cert: PeerCertificate): Error|undefined => {
-              return checkServerIdentity(sslTargetNameOverride, cert);
-            };
+        const sslTargetNameOverride = this.options[
+          'grpc.ssl_target_name_override'
+        ]!;
+        connectionOptions.checkServerIdentity = (
+          host: string,
+          cert: PeerCertificate
+        ): Error | undefined => {
+          return checkServerIdentity(sslTargetNameOverride, cert);
+        };
         connectionOptions.servername = sslTargetNameOverride;
       }
     }
     const subChannel: Http2SubChannel = new Http2SubChannel(
-        this.target, connectionOptions, this.userAgent, this.options);
+      this.target,
+      connectionOptions,
+      this.userAgent,
+      this.options
+    );
     this.subChannel = subChannel;
     const now = new Date();
     const connectionTimeout: number = Math.max(
-        this.currentBackoffDeadline.getTime() - now.getTime(),
-        MIN_CONNECT_TIMEOUT_MS);
+      this.currentBackoffDeadline.getTime() - now.getTime(),
+      MIN_CONNECT_TIMEOUT_MS
+    );
     const connectionTimerId: NodeJS.Timer = setTimeout(() => {
       // This should trigger the 'close' event, which will send us back to
       // TRANSIENT_FAILURE
@@ -211,7 +258,9 @@ export class Http2Channel extends EventEmitter implements Channel {
       // Connection succeeded
       clearTimeout(connectionTimerId);
       this.transitionToState(
-          [ConnectivityState.CONNECTING], ConnectivityState.READY);
+        [ConnectivityState.CONNECTING],
+        ConnectivityState.READY
+      );
     };
     subChannel.once('connect', this.subChannelConnectCallback);
     this.subChannelCloseCallback = () => {
@@ -220,21 +269,25 @@ export class Http2Channel extends EventEmitter implements Channel {
       /* TODO(murgatroid99): verify that this works for
        * CONNECTING->TRANSITIVE_FAILURE see nodejs/node#16645 */
       this.transitionToState(
-          [ConnectivityState.CONNECTING, ConnectivityState.READY],
-          ConnectivityState.TRANSIENT_FAILURE);
+        [ConnectivityState.CONNECTING, ConnectivityState.READY],
+        ConnectivityState.TRANSIENT_FAILURE
+      );
     };
     subChannel.once('close', this.subChannelCloseCallback);
   }
 
   constructor(
-      address: string, readonly credentials: ChannelCredentials,
-      private readonly options: Partial<ChannelOptions>) {
+    address: string,
+    readonly credentials: ChannelCredentials,
+    private readonly options: Partial<ChannelOptions>
+  ) {
     super();
     for (const option in options) {
       if (options.hasOwnProperty(option)) {
         if (!recognizedOptions.hasOwnProperty(option)) {
           console.warn(
-              `Unrecognized channel argument '${option}' will be ignored.`);
+            `Unrecognized channel argument '${option}' will be ignored.`
+          );
         }
       }
     }
@@ -250,8 +303,10 @@ export class Http2Channel extends EventEmitter implements Channel {
       this.defaultAuthority = this.target.host;
     }
     this.filterStackFactory = new FilterStackFactory([
-      new CallCredentialsFilterFactory(this), new DeadlineFilterFactory(this),
-      new MetadataStatusFilterFactory(this), new CompressionFilterFactory(this)
+      new CallCredentialsFilterFactory(this),
+      new DeadlineFilterFactory(this),
+      new MetadataStatusFilterFactory(this),
+      new CompressionFilterFactory(this),
     ]);
     this.currentBackoffDeadline = new Date();
     /* The only purpose of these lines is to ensure that this.backoffTimerId has
@@ -260,61 +315,75 @@ export class Http2Channel extends EventEmitter implements Channel {
 
     // Build user-agent string.
     this.userAgent = [
-      options['grpc.primary_user_agent'], `grpc-node-js/${clientVersion}`,
-      options['grpc.secondary_user_agent']
-    ].filter(e => e).join(' ');  // remove falsey values first
+      options['grpc.primary_user_agent'],
+      `grpc-node-js/${clientVersion}`,
+      options['grpc.secondary_user_agent'],
+    ]
+      .filter(e => e)
+      .join(' '); // remove falsey values first
   }
 
   _startHttp2Stream(
-      authority: string, methodName: string, stream: Http2CallStream,
-      metadata: Metadata) {
-    const finalMetadata: Promise<Metadata> =
-        stream.filterStack.sendMetadata(Promise.resolve(metadata.clone()));
+    authority: string,
+    methodName: string,
+    stream: Http2CallStream,
+    metadata: Metadata
+  ) {
+    const finalMetadata: Promise<Metadata> = stream.filterStack.sendMetadata(
+      Promise.resolve(metadata.clone())
+    );
     Promise.all([finalMetadata, this.connect()])
-        .then(([metadataValue]) => {
-          const headers = metadataValue.toHttp2Headers();
-          headers[HTTP2_HEADER_AUTHORITY] = authority;
-          headers[HTTP2_HEADER_USER_AGENT] = this.userAgent;
-          headers[HTTP2_HEADER_CONTENT_TYPE] = 'application/grpc';
-          headers[HTTP2_HEADER_METHOD] = 'POST';
-          headers[HTTP2_HEADER_PATH] = methodName;
-          headers[HTTP2_HEADER_TE] = 'trailers';
-          if (this.connectivityState === ConnectivityState.READY) {
-            const subChannel: Http2SubChannel = this.subChannel!;
-            subChannel.startCallStream(metadataValue, stream);
-          } else {
-            /* In this case, we lost the connection while finalizing
-             * metadata. That should be very unusual */
-            setImmediate(() => {
-              this._startHttp2Stream(authority, methodName, stream, metadata);
-            });
-          }
-        })
-        .catch((error: Error&{code: number}) => {
-          // We assume the error code isn't 0 (Status.OK)
-          stream.cancelWithStatus(
-              error.code || Status.UNKNOWN,
-              `Getting metadata from plugin failed with error: ${
-                  error.message}`);
-        });
+      .then(([metadataValue]) => {
+        const headers = metadataValue.toHttp2Headers();
+        headers[HTTP2_HEADER_AUTHORITY] = authority;
+        headers[HTTP2_HEADER_USER_AGENT] = this.userAgent;
+        headers[HTTP2_HEADER_CONTENT_TYPE] = 'application/grpc';
+        headers[HTTP2_HEADER_METHOD] = 'POST';
+        headers[HTTP2_HEADER_PATH] = methodName;
+        headers[HTTP2_HEADER_TE] = 'trailers';
+        if (this.connectivityState === ConnectivityState.READY) {
+          const subChannel: Http2SubChannel = this.subChannel!;
+          subChannel.startCallStream(metadataValue, stream);
+        } else {
+          /* In this case, we lost the connection while finalizing
+           * metadata. That should be very unusual */
+          setImmediate(() => {
+            this._startHttp2Stream(authority, methodName, stream, metadata);
+          });
+        }
+      })
+      .catch((error: Error & { code: number }) => {
+        // We assume the error code isn't 0 (Status.OK)
+        stream.cancelWithStatus(
+          error.code || Status.UNKNOWN,
+          `Getting metadata from plugin failed with error: ${error.message}`
+        );
+      });
   }
 
   createCall(
-      method: string, deadline: Deadline|null|undefined,
-      host: string|null|undefined, parentCall: Call|null|undefined,
-      propagateFlags: number|null|undefined): Call {
+    method: string,
+    deadline: Deadline | null | undefined,
+    host: string | null | undefined,
+    parentCall: Call | null | undefined,
+    propagateFlags: number | null | undefined
+  ): Call {
     if (this.connectivityState === ConnectivityState.SHUTDOWN) {
       throw new Error('Channel has been shut down');
     }
     const finalOptions: CallStreamOptions = {
-      deadline: (deadline === null || deadline === undefined) ? Infinity :
-                                                                deadline,
+      deadline:
+        deadline === null || deadline === undefined ? Infinity : deadline,
       flags: propagateFlags || 0,
       host: host || this.defaultAuthority,
-      parentCall: parentCall || null
+      parentCall: parentCall || null,
     };
     const stream: Http2CallStream = new Http2CallStream(
-        method, this, finalOptions, this.filterStackFactory);
+      method,
+      this,
+      finalOptions,
+      this.filterStackFactory
+    );
     return stream;
   }
 
@@ -334,7 +403,9 @@ export class Http2Channel extends EventEmitter implements Channel {
       if (!this.connecting) {
         this.connecting = new Promise((resolve, reject) => {
           this.transitionToState(
-              [ConnectivityState.IDLE], ConnectivityState.CONNECTING);
+            [ConnectivityState.IDLE],
+            ConnectivityState.CONNECTING
+          );
           const onConnect = () => {
             this.connecting = null;
             this.removeListener('shutdown', onShutdown);
@@ -356,14 +427,18 @@ export class Http2Channel extends EventEmitter implements Channel {
   getConnectivityState(tryToConnect: boolean): ConnectivityState {
     if (tryToConnect) {
       this.transitionToState(
-          [ConnectivityState.IDLE], ConnectivityState.CONNECTING);
+        [ConnectivityState.IDLE],
+        ConnectivityState.CONNECTING
+      );
     }
     return this.connectivityState;
   }
 
   watchConnectivityState(
-      currentState: ConnectivityState, deadline: Date|number,
-      callback: (error?: Error) => void) {
+    currentState: ConnectivityState,
+    deadline: Date | number,
+    callback: (error?: Error) => void
+  ) {
     if (this.connectivityState !== currentState) {
       /* If the connectivity state is different from the provided currentState,
        * we assume that a state change has successfully occurred */
@@ -400,10 +475,13 @@ export class Http2Channel extends EventEmitter implements Channel {
       throw new Error('Channel has been shut down');
     }
     this.transitionToState(
-        [
-          ConnectivityState.CONNECTING, ConnectivityState.READY,
-          ConnectivityState.TRANSIENT_FAILURE, ConnectivityState.IDLE
-        ],
-        ConnectivityState.SHUTDOWN);
+      [
+        ConnectivityState.CONNECTING,
+        ConnectivityState.READY,
+        ConnectivityState.TRANSIENT_FAILURE,
+        ConnectivityState.IDLE,
+      ],
+      ConnectivityState.SHUTDOWN
+    );
   }
 }
