@@ -1,9 +1,26 @@
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
 import * as zlib from 'zlib';
 
-import {Call, WriteFlags, WriteObject} from './call-stream';
-import {Channel} from './channel';
-import {BaseFilter, Filter, FilterFactory} from './filter';
-import {Metadata, MetadataValue} from './metadata';
+import { Call, WriteFlags, WriteObject } from './call-stream';
+import { Channel } from './channel';
+import { BaseFilter, Filter, FilterFactory } from './filter';
+import { Metadata, MetadataValue } from './metadata';
 
 abstract class CompressionHandler {
   protected abstract compressMessage(message: Buffer): Promise<Buffer>;
@@ -54,8 +71,11 @@ class IdentityHandler extends CompressionHandler {
   }
 
   decompressMessage(message: Buffer): Promise<Buffer> {
-    return Promise.reject<Buffer>(new Error(
-        'Received compressed message but "grpc-encoding" header was identity'));
+    return Promise.reject<Buffer>(
+      new Error(
+        'Received compressed message but "grpc-encoding" header was identity'
+      )
+    );
   }
 }
 
@@ -116,15 +136,20 @@ class UnknownHandler extends CompressionHandler {
     super();
   }
   compressMessage(message: Buffer): Promise<Buffer> {
-    return Promise.reject<Buffer>(new Error(
+    return Promise.reject<Buffer>(
+      new Error(
         `Received message compressed wth unsupported compression method ${
-            this.compressionName}`));
+          this.compressionName
+        }`
+      )
+    );
   }
 
   decompressMessage(message: Buffer): Promise<Buffer> {
     // This should be unreachable
     return Promise.reject<Buffer>(
-        new Error(`Compression method not supported: ${this.compressionName}`));
+      new Error(`Compression method not supported: ${this.compressionName}`)
+    );
   }
 }
 
@@ -170,13 +195,16 @@ export class CompressionFilter extends BaseFilter implements Filter {
      * and the output is a framed and possibly compressed message. For this
      * reason, this filter should be at the bottom of the filter stack */
     const resolvedMessage: WriteObject = await message;
-    const compress = resolvedMessage.flags === undefined ?
-        false :
-        (resolvedMessage.flags & WriteFlags.NoCompress) === 0;
+    const compress =
+      resolvedMessage.flags === undefined
+        ? false
+        : (resolvedMessage.flags & WriteFlags.NoCompress) === 0;
     return {
       message: await this.sendCompression.writeMessage(
-          resolvedMessage.message, compress),
-      flags: resolvedMessage.flags
+        resolvedMessage.message,
+        compress
+      ),
+      flags: resolvedMessage.flags,
     };
   }
 
@@ -189,8 +217,8 @@ export class CompressionFilter extends BaseFilter implements Filter {
   }
 }
 
-export class CompressionFilterFactory implements
-    FilterFactory<CompressionFilter> {
+export class CompressionFilterFactory
+  implements FilterFactory<CompressionFilter> {
   constructor(private readonly channel: Channel) {}
   createFilter(callStream: Call): CompressionFilter {
     return new CompressionFilter();
