@@ -1,3 +1,20 @@
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
 export enum Status {
   OK = 0,
   CANCELLED,
@@ -15,11 +32,11 @@ export enum Status {
   INTERNAL,
   UNAVAILABLE,
   DATA_LOSS,
-  UNAUTHENTICATED
+  UNAUTHENTICATED,
 }
 
 export enum LogVerbosity {
   DEBUG = 0,
   INFO,
-  ERROR
+  ERROR,
 }
