@@ -1,10 +1,31 @@
-import {ChannelCredentials} from './channel-credentials';
-import {ChannelOptions} from './channel-options';
-import {Client} from './client';
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
 
-export interface Serialize<T> { (value: T): Buffer; }
+import { ChannelCredentials } from './channel-credentials';
+import { ChannelOptions } from './channel-options';
+import { Client } from './client';
 
-export interface Deserialize<T> { (bytes: Buffer): T; }
+export interface Serialize<T> {
+  (value: T): Buffer;
+}
+
+export interface Deserialize<T> {
+  (bytes: Buffer): T;
+}
 
 export interface MethodDefinition<RequestType, ResponseType> {
   path: string;
@@ -28,7 +49,7 @@ export interface ProtobufTypeDefinition {
 }
 
 export interface PackageDefinition {
-  [index: string]: ServiceDefinition|ProtobufTypeDefinition;
+  [index: string]: ServiceDefinition | ProtobufTypeDefinition;
 }
 
 /**
@@ -40,7 +61,7 @@ const requesterFuncs = {
   unary: Client.prototype.makeUnaryRequest,
   server_stream: Client.prototype.makeServerStreamRequest,
   client_stream: Client.prototype.makeClientStreamRequest,
-  bidi: Client.prototype.makeBidiStreamRequest
+  bidi: Client.prototype.makeBidiStreamRequest,
 };
 
 export interface ServiceClient extends Client {
@@ -48,8 +69,11 @@ export interface ServiceClient extends Client {
 }
 
 export interface ServiceClientConstructor {
-  new(address: string, credentials: ChannelCredentials,
-      options?: Partial<ChannelOptions>): ServiceClient;
+  new (
+    address: string,
+    credentials: ChannelCredentials,
+    options?: Partial<ChannelOptions>
+  ): ServiceClient;
   service: ServiceDefinition;
 }
 
@@ -68,8 +92,10 @@ export interface ServiceClientConstructor {
  *     {@link grpc.Client}, and has the same arguments as that constructor.
  */
 export function makeClientConstructor(
-    methods: ServiceDefinition, serviceName: string,
-    classOptions?: {}): ServiceClientConstructor {
+  methods: ServiceDefinition,
+  serviceName: string,
+  classOptions?: {}
+): ServiceClientConstructor {
   if (!classOptions) {
     classOptions = {};
   }
@@ -79,7 +105,7 @@ export function makeClientConstructor(
     [methodName: string]: Function;
   }
 
-  Object.keys(methods).forEach((name) => {
+  Object.keys(methods).forEach(name => {
     const attrs = methods[name];
     let methodType: keyof typeof requesterFuncs;
     // TODO(murgatroid99): Verify that we don't need this anymore
@@ -101,14 +127,18 @@ export function makeClientConstructor(
     }
     const serialize = attrs.requestSerialize;
     const deserialize = attrs.responseDeserialize;
-    const methodFunc =
-        partial(requesterFuncs[methodType], attrs.path, serialize, deserialize);
+    const methodFunc = partial(
+      requesterFuncs[methodType],
+      attrs.path,
+      serialize,
+      deserialize
+    );
     ServiceClientImpl.prototype[name] = methodFunc;
     // Associate all provided attributes with the method
     Object.assign(ServiceClientImpl.prototype[name], attrs);
     if (attrs.originalName) {
       ServiceClientImpl.prototype[attrs.originalName] =
-          ServiceClientImpl.prototype[name];
+        ServiceClientImpl.prototype[name];
     }
   });
 
@@ -118,21 +148,27 @@ export function makeClientConstructor(
 }
 
 function partial(
-    fn: Function, path: string, serialize: Function,
-    deserialize: Function): Function {
+  fn: Function,
+  path: string,
+  serialize: Function,
+  deserialize: Function
+): Function {
   // tslint:disable-next-line:no-any
   return function(this: any, ...args: any[]) {
     return fn.call(this, path, serialize, deserialize, ...args);
   };
 }
 
-export type GrpcObject = {
-  [index: string]: GrpcObject|ServiceClientConstructor|ProtobufTypeDefinition;
-};
+export interface GrpcObject {
+  [index: string]:
+    | GrpcObject
+    | ServiceClientConstructor
+    | ProtobufTypeDefinition;
+}
 
 function isProtobufTypeDefinition(
-    obj: ServiceDefinition|
-    ProtobufTypeDefinition): obj is ProtobufTypeDefinition {
+  obj: ServiceDefinition | ProtobufTypeDefinition
+): obj is ProtobufTypeDefinition {
   return 'format' in obj;
 }
 
@@ -141,8 +177,9 @@ function isProtobufTypeDefinition(
  * @param packageDef The package definition object.
  * @return The resulting gRPC object.
  */
-export function loadPackageDefinition(packageDef: PackageDefinition):
-    GrpcObject {
+export function loadPackageDefinition(
+  packageDef: PackageDefinition
+): GrpcObject {
   const result: GrpcObject = {};
   for (const serviceFqn in packageDef) {
     if (packageDef.hasOwnProperty(serviceFqn)) {
