@@ -1,26 +1,44 @@
-import {Duplex, Readable, Writable} from 'stream';
-import {EmitterAugmentation1} from './events';
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import { Duplex, Readable, Writable } from 'stream';
+import { EmitterAugmentation1 } from './events';
 
 // tslint:disable:no-any
 
-export type WriteCallback = (error: Error|null|undefined) => void;
+export type WriteCallback = (error: Error | null | undefined) => void;
 
 export interface IntermediateObjectReadable<T> extends Readable {
-  read(size?: number): any&T;
+  read(size?: number): any & T;
 }
 
 export type ObjectReadable<T> = {
   read(size?: number): T;
-}&EmitterAugmentation1<'data', T>&IntermediateObjectReadable<T>;
+} & EmitterAugmentation1<'data', T> &
+  IntermediateObjectReadable<T>;
 
 export interface IntermediateObjectWritable<T> extends Writable {
-  _write(chunk: any&T, encoding: string, callback: Function): void;
-  write(chunk: any&T, cb?: WriteCallback): boolean;
-  write(chunk: any&T, encoding?: any, cb?: WriteCallback): boolean;
+  _write(chunk: any & T, encoding: string, callback: Function): void;
+  write(chunk: any & T, cb?: WriteCallback): boolean;
+  write(chunk: any & T, encoding?: any, cb?: WriteCallback): boolean;
   setDefaultEncoding(encoding: string): this;
   end(): void;
-  end(chunk: any&T, cb?: Function): void;
-  end(chunk: any&T, encoding?: any, cb?: Function): void;
+  end(chunk: any & T, cb?: Function): void;
+  end(chunk: any & T, encoding?: any, cb?: Function): void;
 }
 
 export interface ObjectWritable<T> extends IntermediateObjectWritable<T> {
@@ -42,4 +60,6 @@ export type ObjectDuplex<T, U> = {
   end(): void;
   end(chunk: T, cb?: Function): void;
   end(chunk: T, encoding?: any, cb?: Function): void;
-}&Duplex&ObjectWritable<T>&ObjectReadable<U>;
+} & Duplex &
+  ObjectWritable<T> &
+  ObjectReadable<U>;
