@@ -1,14 +1,31 @@
-import {StatusObject} from './call-stream';
-import {Status} from './constants';
-import {Metadata} from './metadata';
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import { StatusObject } from './call-stream';
+import { Status } from './constants';
+import { Metadata } from './metadata';
 
 /**
  * A builder for gRPC status objects.
  */
 export class StatusBuilder {
-  private code: Status|null;
-  private details: string|null;
-  private metadata: Metadata|null;
+  private code: Status | null;
+  private details: string | null;
+  private metadata: Metadata | null;
 
   constructor() {
     this.code = null;
