@@ -1,4 +1,32 @@
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import * as loader from '@grpc/proto-loader';
 import * as assert from 'assert';
+
+import { GrpcObject, loadPackageDefinition } from '../src/make-client';
+
+const protoLoaderOptions = {
+  keepCase: true,
+  longs: String,
+  enums: String,
+  defaults: true,
+  oneofs: true,
+};
 
 export function mockFunction(): never {
   throw new Error('Not implemented');
@@ -21,7 +49,7 @@ export namespace assert2 {
       assert.throws(() => {
         throw e;
       });
-      throw e;  // for type safety only
+      throw e; // for type safety only
     }
   }
 
@@ -31,7 +59,7 @@ export namespace assert2 {
    */
   function mustCallsSatisfied(): boolean {
     let result = true;
-    toCall.forEach((value) => {
+    toCall.forEach(value => {
       result = result && value === 0;
     });
     return result;
@@ -46,8 +74,9 @@ export namespace assert2 {
    * @param fn The function to wrap.
    */
   // tslint:disable:no-any
-  export function mustCall<T>(fn: (...args: any[]) => T):
-      (...args: any[]) => T {
+  export function mustCall<T>(
+    fn: (...args: any[]) => T
+  ): (...args: any[]) => T {
     const existingValue = toCall.get(fn);
     if (existingValue !== undefined) {
       toCall.set(fn, existingValue + 1);
@@ -82,4 +111,9 @@ export namespace assert2 {
       fn();
     }
   }
+}
+
+export function loadProtoFile(file: string): GrpcObject {
+  const packageDefinition = loader.loadSync(file, protoLoaderOptions);
+  return loadPackageDefinition(packageDefinition);
 }
