@@ -1,3 +1,20 @@
+/*
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
 import * as assert from 'assert';
 
 import * as grpc from '../src';
@@ -22,11 +39,11 @@ describe('Logging', () => {
   });
 
   it('gates logging based on severity', () => {
-    const output: Array<string|string[]> = [];
+    const output: Array<string | string[]> = [];
     const logger: Partial<Console> = {
       error(...args: string[]): void {
         output.push(args);
-      }
+      },
     };
 
     logging.setLogger(logger);
@@ -48,8 +65,13 @@ describe('Logging', () => {
     logging.log(grpc.logVerbosity.INFO, 7, 8);
     logging.log(grpc.logVerbosity.ERROR, 'j', 'k');
 
-    assert.deepStrictEqual(
-        output,
-        [['a', 'b', 'c'], ['d', 'e'], ['f'], ['g'], ['h', 'i'], ['j', 'k']]);
+    assert.deepStrictEqual(output, [
+      ['a', 'b', 'c'],
+      ['d', 'e'],
+      ['f'],
+      ['g'],
+      ['h', 'i'],
+      ['j', 'k'],
+    ]);
   });
 });
